(dev/core#217) PrevNext - Remove references to entity_table and entity_id2 from service
The PrevNext service/interface aims to be a replaceable component for use
the search-caching (but not deduping).

The interface that we produced from refactoring includes several references
to `entity_table` and `entity_id2` -- these values are part of the SQL
table, and they're needed for dedupe, but they don't seem to convey anything
meaningful for search-caching.  Including these fields makes the interface
more complicated -- which will make it hard to implement other variants.

The general gist of this commit is that we no longer fill those two columns,
and we no longer read them.

In a couple functions, we split the new OOP implementation
(`CRM_Core_PrevNextCache_Sql`) from the traditional static BAO
implementation (`CRM_Core_BAO_PrevNext`) so that we can omit these fields.