Start building export processor class.
A fundamental cause of code mess in the Export class is that it is a mishmash of static functions with variables passed around crazily. This starts the process of moving work to a object oriented class.

I think this migratory approach is the best plan to clean up the code

I did move some vars to being accessed via 'self' but there is leakage across tests so going
full OOO seems like a better approach