Fix export of location types where name does not match label.
There is an issue with the code using the label and name of location types (and other
things but not in scope at the moment) indiscriminately.

This change switches to keying by ID rather than label/name and supports
it in the query object. The test cover is pretty intensive here.