Remove net_amount from the Back office contribution form.
Net amount is calculated in the BAO if not set so asking users to enter it does not gain us anything.

Currently we require them to calculate it & edit it if they change the fee_amount or total_amount
and we validate their data entry, and make them re-do it if they get it wrong.

Since the field is pretty hidden it's unintuitive to need to change it and
makes for a painful contribution update process.

The current behaviour dates back to svn days - however since 2015 the BAO has handled the possibility
of it not being set.

In addition we have had code issues with the comparison around currency & float comparison issues
eg. https://github.com/civicrm/civicrm-core/pull/11485
and https://lab.civicrm.org/dev/core/issues/260 ( in the latter case the data saves
correctly without net_amount and incorrectly if it is changed to meet the form validation rule.
(There was a proposal in Mar 2017 to address that https://github.com/civicrm/civicrm-core/pull/9948#issuecomment-285040403 by fixing the calculation but I believe just dropping the field is better).

Our unit tests test the form submissions but for some reason net_amoutn was removed from the tests
https://github.com/civicrm/civicrm-core/pull/9948/files#diff-40e2e0f106ba620465acf3a9a81f2498L1535
meaning our test coverage is more reliable without it being set.