(dev/core#217) Cleanup PrevNext interface. Move Query-BAO hackery back to its old place.
The general context of this code is roughly as follows: "We've already
filled up the prevnext cache with a bunch of contact-IDs.  The user wants to
view a page of 50 contacts. The prevnext cache has all the contact-IDs,
but we want to lookup full information about the specific 50 contacts
on this page."

The `CRM_Contact_BAO_Query` is good for looking up the "full information"
some contacts, but it doesn't know how to filter based on the prevnext
cache. One needs a way to add that bit of information.

Before
------

* The contract for `CRM_Core_PrevNextCache_Interface::fetch()` was extremely
  nuanced -- it basically requires you to call the Query BAO and hack
  its SQL output.
* `CRM_Core_PrevNextCache_Sql::fetch()` (and its equivalent predecessor
  `CRM_Contact_BAO_Query::getCachedContacts()`) achieved this by
   splicing some extra clause into the SQL query -- specifically,
   a `JOIN` on the `civicrm_prevnext_cache` and some
   `GROUP BY`/`ORDER BY`/`LIMIT`.
* `CRM_Core_PrevNextCache_Redis::fetch()` achieved this with two
  steps -- first, reading the contact-list from the cache; and then
  splicing an extra `WHERE` into the SQL query.
* There is not unit test coverage for `::fetch()`.

After
-----

* The contract for `CRM_Core_PrevNextCache_Interface::fetch()` is
  much simpler - you don't need to know anything about the query BAO.
  You just return one page worth of ID's from the cache.
* The SQL manipulation has moved back to its original place in
  `CRM_Contact_BAO_Query::getCachedContacts()`. However, we
  do the same manipulation regardless of cache-backend -- ie
  we always splice in the `WHERE` (and not the `JOIN`).
* There is unit test coverage for `::fetch()`.