dev/core#316 - Update long string cache key test results
This is replaces 95b65673b6259be3a02ed6510e7562939d35b6c8.  The intent is to
have a delimiter every ten chars so that it's easy to read/confirm the
length of the string.

The delimiter was space, but this became a longer encoded char (`-20`)
and threw off the numbers. Switching to dash just gives a different encoded char.
To get the counts right, it needs ot be a pass-through char... like underscore.