(NFC) Upgrade - Improve example of an upgrade message for a new permission
@bgm pointed out that the permission name should be translated as well.
This left me wondering...  why do we bother with parameterizing the string
with a `%1` if it's all going to be translated anyway?

There is a plausible reason.  It's not super-critical, but it's plausible --
the permission name is used in more contexts (like the permission mgmt
screen in D7/WP *as well as* the upgrade screen).  This arrangement ensures
that the name appears the same in *all those contexts*, even in various
edge-cases of the translation lifecycle (e.g.  where the permission-names
have been translated but the upgrade-messages have not).  Again...  it's not
a big deal, but it is a slightly better default.