Fix duplicate merge to not disregard zero values.
Currently a field with 0 in it will not be carried across in a merge. In fact in a checkbox situation the differenct between
0 and null is significant (I chose not to opt in vs
I was never presented with a choice).

This patch preserves the 0. I would note that
further digging is required to see how a confict is handled
- ie. if one contact has 0 & the other has 1 is this blocked
as a confligct

Additional test on conflict handling