Partial revert of pull/12559
https://github.com/civicrm/civicrm-core/pull/12559/files is currently in RC.

While we find that this change improves performance and we are running it in production
we have identified that performance is very poor in the specific scenario where only
email is linked to 39000 assignees - in this case it was an old experimental civimail
activity. In the interests of prudence I am reverting the parts that switch function
in the rc.