Remove empty space above menu
CiviCRM doesn't use the SubHead Toolbar in Joomla's 3.x admin theme, leaving an empty region. Fix removes that, full-widths the content and adds padding for consistency. Tested with and without Shoreditch. Also includes a specific change to deal with Shoreditch that create a regression without. Unfortunate use of '!important' to override one hardcoded element and the order of file loading.