Remove use of LOWER from city & street searches and fix international strings test.
Adding LOWER to mysql queries makes them slower. lowercasing php strings
breaks under some character sets with some server configs.

In trying to review PR 12364 I dug into the strotolower behaviour. I found that
in the path altered in this PR we are setting the value to lower case but not the
mysql clause. I ran through ContactTest without these lines & found it
only hit this strtolower in one test - testInternationalStrings which searches
by organization name. I found the this test actually failed for me
when I ran it locally - but it passed when I removed the strotlower lines.

Wonder in jenkins finds the same