Prototype for metadata based fields on search screens.
This is a first step towards making search screens more metadata driven. It switches the activity_type_id field
to being metadata-added and adds support for activity_type_id=1 via the url.

The setting of defaults is more like 'normal' forms where they are loaded & can be changed in the UI
rather than 'search methods' where they only work in conjunction with 'force'
(which also still works).

This comes out of my concern that
a) we already have url defaults added badly on this form &
b) we have had more than one PR to add more defaults - eg. https://github.com/civicrm/civicrm-core/pull/12455

So I think we should define how we really want entity data to be used with search forms & move towards it.

Note also that the tpls for Contribution Search and one of the contact search panes have been amended to
be metadata driven. I haven't tried to align with that in this step but it should be down the track