Fix misplaced hyphens in regex
Inside a character range surrounded by [], the hyphen is interpreted
as defining a range, unless it's the first character. Move the
hyphens to the first character in the square brackets where it's
supposed to be interpreted as a hyphen.

Under PHP 7.3, the previous code failed with error message
preg_match(): Compilation failed: invalid range in character class at offset 7