Declare 'amount' as required on ContributionRecur api
This is effectively required already as you get a DB error without it.

The field is not required at the DB level (it probably should be that too) but
it has no default value so you wind up with

[nativecode=1364 ** Field 'amount' doesn't have a default value]

Note that making it required at the DB level won't change the need for this
patch & is separate. The api declares what is required at the api level
- somethings are required at the DB level but not at the api level as they are
calculated in the BAO