Add Case Manager Logic for closed cases (inactive case roles)
In the getCaseRoles function, added is_active and end_date to the query and to the values array (for use in the case roles table on the manage case screen controlled in the caseview.tpl)

In the getCaseManagerContact > managerRoleQuery function, I removed the portion of the where clause that was only returning active case managers. I added the is_active and end_date to the query here as well. Then, I added logic that first searched for an active case manager, if none found then it would show the most recently 'ended' manager. (This controls what name displays on case listings, like on a contact card or case search. This way for closed cases a case manager is still listed.)