Add is_active and end_date to caseRelationships array
This builds on @jitendrapurohit pull https://github.com/civicrm/civicrm-core/pull/13144/commits/a56aac79df8451e5d22d4c061de6a7f5a84907e7

First, set activeOnly to false in the getCaseRoles call - to ensure we get both open and inactive roles.

Add the is_active and end_date as rows to the $caseRoles array.

Then using those rows - add to the data table. Use is_active to add the 'disabled' class where appropriate. Add a column to display the end date. (If an end date is present, do not show the actions.)

To help avoid confusion, swap the trash can icon for the ban icon and change label text to 'expire'. (As that action isn't 'trashing' anything, it just adds an end date to the relationship - so the relationship still shows in the list.)