Add columns to data table and show inactive checkbox
In the case roles data table, add a column for end_date. 

Add a smarty checkbox called 'show inactive relationships.' JS to show/hide rows with the 'disabled' class based on clicking the checkbox. Additional JS logic checks the box by default on closed cases.

Add to the ajax call, an 'on complete' to hide 'disabled' rows if the case is open. If there are no disabled rows to hide, gray out the 'show inactive relationships' checkbox and text.

Adjust DeleteCaseRoleDialog text to be less consistent (change to 'expire').