dev/core#690 - Civi\API - Add a check on entity_table existing
This adds a check in DynamicFKAuthorization to verify that the
entity_table in the API request actually exists as a table.

Additionally, this changes a test case in api_v3_AttachmentTest to
enable permission checking. This is necessary because the change to
DynamicFKAuthorization means trusted API calls can now attach files
to *any* entity unless check_permissions is set, in which case it's
only possible for allowed delegates.